Refactoring ied-list
